+++ packages/core/src/index.ts
@@ -1,6 +1,7 @@
 import Config from './Config/Config';
 import { IConfig } from './Config/types';
 import Html from './Html/Html';
+import { IScriptConfig, IHTMLTag } from './Html/types';
 import Route from './Route/Route';
 import { IRoute } from './Route/types';
 import Service, { IServiceOpts } from './Service/Service';
@@ -9,5 +10,5 @@ import UmiError from './Logger/UmiError';
 import Logger from './Logger/Logger';
 
 export { Config, Html, Route, Service, PluginAPI };
-export { IRoute, IConfig, IServiceOpts };
+export { IRoute, IConfig, IServiceOpts, IScriptConfig, IHTMLTag };
 export { Logger, UmiError };
